# [Enhancement] window function binary search optimization