## What type of PR is this：
- [ ] bug
- [ ] feature
- [x] enhancement
- [ ] refactor
- [ ] others

## Which issues of this PR fixes ：
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #7904

## Enhancement

Binary search will be replaced with loop search when cardinality of partition is high. But how to determine whether the cardinality is high or not? 

**Overhead analysis**

* Binary Search: `O(N/S logN)`, where `S` denote the average partition size
* Loop Search：`O(N)`

The default `chunk_size` is set to `4096`, let `N = 4096`, the overhead of single binary search is `log 4096 = 12`。

So, theoretically, regardless of compiler optimizations, binary search is better when `S > 12` , otherwise Loop Search is better. And considering the error of estimation and for simplicity, we set the threshould to `8`, blow which is considered as high cardinality.

## Test

* 1fe/3be, 64c/128g

```sql
-- Q1, ssb-100g, dop=8
SELECT max(wx) FROM ( SELECT avg(LO_TAX) OVER (PARTITION BY LO_ORDERKEY,LO_PARTKEY ORDER BY LO_TAX ROWS BETWEEN 1 PRECEDING AND 1 FOLLOWING) AS wx FROM lineorder_flat )a;

-- Q2, ssb-1g, dop=1
SELECT max(wx) FROM (
SELECT avg(LO_TAX) OVER (PARTITION BY C_REGION,C_NATION,C_CITY,S_CITY,S_NATION,S_REGION,P_TYPE ORDER BY LO_ORDERKEY ROWS BETWEEN 1 PRECEDING AND 1 FOLLOWING) AS wx FROM lineorder_flat
)a;

-- Q3(same as Q2), ssb-1g, dop=8
SELECT max(wx) FROM (
SELECT avg(LO_TAX) OVER (PARTITION BY C_REGION,C_NATION,C_CITY,S_CITY,S_NATION,S_REGION,P_TYPE ORDER BY LO_ORDERKEY ROWS BETWEEN 1 PRECEDING AND 1 FOLLOWING) AS wx FROM lineorder_flat
)a;
```

|  | query time | analytic time | partition search time |
|:--|:--|:--|:--|
| Q1-Before | 8.219s | 3s899ms | 733.219ms |
| Q1-After | 5.695s | 2s625ms | 48.535ms |
| Q2-Before | 17.267s | 6s115ms | 3s284ms |
| Q2-After | 13.056s | 2s94ms | 394.552ms |
| Q3-Before | 2.199s | 773.938ms | 385.516ms |
| Q3-After | 1.770s | 279.774ms | 49.531ms |